[ 🔨 refactor] Todo와 Comment Service 로직 분리
