+++ src/main/kotlin/com/example/mytodo/domain/comment/controller/CommentController.kt
@@ -3,7 +3,7 @@ package com.example.mytodo.domain.comment.controller
 import com.example.mytodo.domain.comment.dto.CommentCreateRequestDto
 import com.example.mytodo.domain.comment.dto.CommentResponseDto
 import com.example.mytodo.domain.comment.dto.CommentUpdateRequestDto
-import com.example.mytodo.domain.todo.service.TodoService
+import com.example.mytodo.domain.comment.service.CommentService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping("/todo/{todoId}/comment")
 class CommentController(
-    private val todoService: TodoService
+    private val commentService: CommentService
 ) {
 
     @PostMapping
@@ -20,13 +20,13 @@ class CommentController(
         @PathVariable todoId: Long,
         @RequestBody commentCreateRequestDto: CommentCreateRequestDto)
     :ResponseEntity<CommentResponseDto>{
-        return ResponseEntity.status(HttpStatus.CREATED).body(todoService.createComment(todoId, commentCreateRequestDto))
+        return ResponseEntity.status(HttpStatus.CREATED).body(commentService.createComment(todoId, commentCreateRequestDto))
     }
 
     @GetMapping
     fun getAllComments(@PathVariable todoId: Long)
     :ResponseEntity<List<CommentResponseDto>>{
-        return ResponseEntity.status(HttpStatus.OK).body(todoService.getAllComments(todoId))
+        return ResponseEntity.status(HttpStatus.OK).body(commentService.getAllComments(todoId))
     }
 
     @PutMapping("/{commentId}")
@@ -35,7 +35,7 @@ class CommentController(
         @PathVariable commentId: Long,
         @RequestBody commentUpdateRequestDto: CommentUpdateRequestDto
     ):ResponseEntity<CommentResponseDto>{
-        return ResponseEntity.status(HttpStatus.OK).body(todoService.updateComment(todoId, commentId, commentUpdateRequestDto))
+        return ResponseEntity.status(HttpStatus.OK).body(commentService.updateComment(todoId, commentId, commentUpdateRequestDto))
     }
 
     @DeleteMapping("/{commentId}")
@@ -43,6 +43,7 @@ class CommentController(
         @PathVariable todoId: Long,
         @PathVariable commentId: Long,
         ):ResponseEntity<Unit>{
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).body(todoService.deleteComment(todoId, commentId))
+        commentService.deleteComment(todoId, commentId)
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build()
     }
 }
